requirements.txt
flask
flask-cors
numpy==1.24.3
tensorflow==2.13.0
scikit-learn
spacy
opencv-python
cryptography
uvicorn
asgiref
torch==2.0.1
typing-extensions==4.5.0
pydantic==2.5.2
python-dotenv
pandas==2.0.3

# LangChain ecosystem
langchain-core==0.1.17
langchain==0.0.292
langchain-community==0.0.10
langsmith==0.0.92

# Additional requirements
openai
python-whois
textblob
pytesseract
Pillow
imagehash
tld 